+++ subsys/usb/usb_c/CMakeLists.txt
@@ -2,4 +2,19 @@
 
 zephyr_library()
 
-zephyr_library_sources_ifdef(CONFIG_USBC_STACK usbc_timer.c usbc_stack.c usbc_tc.c usbc_prl.c usbc_pe.c)
+# Common USB-C Stack files
+zephyr_library_sources_ifdef(
+	CONFIG_USBC_STACK
+	usbc_timer.c
+	usbc_stack.c
+	usbc_tc_common.c
+	usbc_pe_common.c
+	usbc_prl.c
+)
+
+# Sink USB-C Stack files
+zephyr_library_sources_ifdef(
+	CONFIG_USBC_CSM_SINK_ONLY
+	usbc_tc_snk_states.c
+	usbc_pe_snk_states.c
+)
