+++ subsys/usb/usb_c/Kconfig
@@ -39,6 +39,17 @@ config BUILD_OUTPUT_VIF
 	help
 	  Generate XML file containing VIF policies during project build.
 
+choice USBC_CSM_TYPE
+	prompt "USB-C Connection State Machine"
+	default USBC_CSM_SINK_ONLY
+
+config USBC_CSM_SINK_ONLY
+	bool "Sink USB-C Connection State Machine"
+	help
+	  Allows the USB-C state machine to function as a Sink
+
+endchoice
+
 module = USBC_STACK
 module-str = usbc stack
 source "subsys/logging/Kconfig.template.log_config"
